FIX! apply was broken
If $op is a builtin, it will return just $ast, when ends the evaluation.

If $op is a lambda, it will return ($ast, $env) which will continue the
evalution TCO-style.

Previously if we a builtin, we would evaluate the return value, which is
wrong. This was causing (list 1) to fail when self-hosting. Because
we'd have (apply list '(1)) which would call (list '(1)), returning (1),
but then evaluate that again, causing a failure.

Oh that was tricky to find...