Fix final self-hosting test failure!
It was this line from tests/step6_file.mal

    (read-string ";; comment")

In normal mode, this would throw a malEmptyInputException, which would be
caught by the system try/catch.

In self-hosting mode, this would be executed within the mal code's try block,
which means our try* special form. I wasn't handling that case in the try/
catch code in the try* special form.