Convert malSeq::m_items from vector to vector*
We still have a lot of double-handling when creating malSequences.
I want to use a vector* instead of a vector so that we can pass the
manually-built vector straight in rather than copying it.

Part one of two.

Convert malSequence::m_items from malObjectVec to malObjectVec*, while
keeping the constructor interfaces the same.