Begin reverse rebase of sdt-cpp
The plan is to start from the completed version, and then work backwards,
removing the features for each step, until we're back at step 0.
Then, rebase that in reverse order so that we get the proper series of
commits.

On the way down, at step N, don't worry about the step < N files. Those will
get forward-ported on the way back up.

This might be just crazy enough to work.